Fix typing issue with schema
The issue is that packaging schema (and form) defines packaging arguments as a list of objects with 'name' and 'value' fields. But the API expects just a JSON object that describes arguments key-value mapping. The transformation happens in processBeforeSubmit function. So it must be called *after* casting to schema, otherwise schema doesn't recognize the type of 'arguments' field.